+++ Source/EventFlow.SQLite/Connections/SQLiteConnection.cs
@@ -21,20 +21,20 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 using EventFlow.Core;
-using EventFlow.Logs;
 using EventFlow.Sql.Connections;
 using EventFlow.SQLite.RetryStrategies;
+using Microsoft.Extensions.Logging;
 
 namespace EventFlow.SQLite.Connections
 {
     public class SQLiteConnection : SqlConnection<ISQLiteConfiguration, ISQLiteErrorRetryStrategy, ISQLiteConnectionFactory>, ISQLiteConnection
     {
         public SQLiteConnection(
-            ILog log,
+            ILogger<SQLiteConnection> logger,
             ISQLiteConfiguration configuration,
             ISQLiteConnectionFactory connectionFactory,
             ITransientFaultHandler<ISQLiteErrorRetryStrategy> transientFaultHandler)
-            : base(log, configuration, connectionFactory, transientFaultHandler)
+            : base(logger, configuration, connectionFactory, transientFaultHandler)
         {
         }
     }